--- pages/seller/index.tsx
@@ -1,9 +1,0 @@
-function SellerPage() {
-  return (
-    <div>
-      <h1>안녕하세요</h1>
-    </div>
-  );
-}
-
-export default SellerPage;
