+++ src/components/AddProduct/ImgSlide.tsx
